e2e tests: improve reliability of login test
Previously the tests were a bit flakey, working some times and not others.  In
this patch we've added the `--runInBand` option to `jest` so that the tests run
sequentially in the same process as the test runny. It may have been that
`jest` was attempting to run the tests in parallel even though the nature of
testing Electron/Spectron doesn't support this and so we would sometimes fail
based on concurrency conflicts.

A few new tests have been added in order to verify that the new updates are
working and to demonstrate how we can start adding testable behaviors into the
app. We still need to improve how the tests are organized and setup to minimize
the noise in the code and isolate the tests.